// src/data/node-types/magazine-issues/types/MagazineIssueNode.ts
export type MagazineIssueNode = {
    id: number
    title: string
    consecutive_number?: number
    issue_number?: number
    issue_year?: number
    release_date?: string
    single_copy_price?: number
    single_copy_price_unit?: string
    pages?: number
    created_at: string
    updated_at: string
}